condensed user logic into seperate app
