+++ server/comment_app/urls.py
@@ -1,5 +1,4 @@
 from django.urls import path
-
 from .views import CommentView
 
 urlpatterns = [
