+++ server/falcon_proj/urls.py
@@ -4,7 +4,8 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path("ticket/", include("ticket_app.urls")),
-    path("signup/", include("signup_app.urls")),
+    # corrected pathing to user
+    path("signup/", include("user_app.urls")),
     path("weather/", include("weather_app.urls")),
     path("event/",include("event_app.urls")),
     path("comment/",include("comment_app.urls"))
