--- server/signup_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SignupAppConfig(AppConfig):
-    name = 'signup_app'
